[refactor] 훅 분리
