+++ src/domains/login/components/LoginRedirectHandler.tsx
@@ -1,68 +1,11 @@
 'use client';
 
-import { useEffect, useState } from 'react';
-import { usePathname, useRouter } from 'next/navigation';
-import { getCookie, removeCookie } from '@/domains/shared/auth/utils/cookie';
-import { useAuthStore } from '@/domains/shared/store/auth';
 import Spinner from '@/shared/components/spinner/Spinner';
 import WelcomeModal from '@/domains/login/components/WelcomeModal';
-import { useToast } from '@/shared/components/toast/useToast';
+import { useLoginRedirect } from '../hooks/useAuthHooks';
 
 function LoginRedirectHandler() {
-  const pathname = usePathname();
-  const router = useRouter();
-  const { user, updateUser } = useAuthStore();
-  const [loading, setLoading] = useState(true);
-  const [welcomeModalOpen, setWelcomeModalOpen] = useState(false);
-  const { toastSuccess } = useToast();
-
-  useEffect(() => {
-    if (!user && loading) {
-      updateUser()
-        .then((fetchedUser) => {
-          if (!fetchedUser) {
-            router.replace('/login');
-          }
-        })
-        .catch(() => {
-          router.replace('/login');
-        })
-        .finally(() => setLoading(false));
-    } else {
-      setLoading(false);
-    }
-  }, [user, loading, updateUser, router]);
-
-  useEffect(() => {
-    if (!user || loading) return;
-
-    const preLoginPath = getCookie('preLoginPath') || '/';
-    // 로그인 상태인데 이전 페이지가 /login이면 메인으로 이동
-    if (user && preLoginPath === '/login') {
-      router.replace('/');
-      removeCookie('preLoginPath');
-      return;
-    }
-
-    // 첫 유저일 경우 모달 오픈
-    if (pathname.startsWith('/login/user/first-user')) {
-      setWelcomeModalOpen(true);
-    }
-    // 기존 유저일 경우
-    else if (pathname.startsWith('/login/user/success')) {
-      toastSuccess(`${user.nickname}님 \n 로그인 성공 🎉`);
-      router.replace(preLoginPath);
-      removeCookie('preLoginPath');
-    }
-  }, [pathname, user, loading, router, toastSuccess]);
-
-  // 환영 모달 닫힐 때 이동
-  const handleCloseWelcomeModal = () => {
-    setWelcomeModalOpen(false);
-    const preLoginPath = getCookie('preLoginPath') || '/';
-    removeCookie('preLoginPath');
-    router.replace(preLoginPath);
-  };
+  const { loading, welcomeModalOpen, handleCloseWelcomeModal, user } = useLoginRedirect();
 
   if (loading) {
     return (
